Created a Bar Plot Visualization for the relationship between experience level and amount of players
Using the data from the players dataset, I made a visualization for the relationship between experience level and the amount of players. The fill in the bar plot signifies subscription, the red area means no, the blue area means yes. My specific question covers the topic of player engagement and how that ties into subscription levels, so I wanted to display that here.