Scatterplot Added
A scatterplot was made to visualize the relationship between age and total hours played. The colours on the datapoints help determine if individuals are subscribed to the newsletter. My specific question focuses on player engagement and how that influences subscription rates, so I wanted to visualize the different variables to see how much they truly influence subscription, and if a pattern could be identified.